Set solidity compiler version to 0.8.6
We were using ^0.8.0 version in Solidity files but it's been increased
to ^0.8.5 as 0.8.4 contains high-severity issue related to the decoder.

Set the solc version in config files to the most recent version
released, which is 0.8.6. The version will have to be pinned in the
particular Solidity module using this library.