Do not reduce allowance in burnFrom for uin256 max allowance
transferFrom does not reduce the allowance if the approval was done for
type(uint256).max value. This approach is implemented in Uniswap LP tokens
and DAI. We also have it in our implementation of ERC20 - it makes sense to
have an explicit way to authorize another contract to spend any amount of
tokens. For example, there could be a treasury contract and a number of
authorized trading contracts taking funds from treasury to make profit in an
automated way. Owner of the treasury contract might want to authorize a fixed
set of trading contracts to take any funds necessary from the treasury so that
the trade (and profit!) can be made.

The same approach as in transferFrom should be applied in burnFrom. First, for
consistency. Secondly, we are eliminating a path when it is possible to use
burnFrom to burn just 1 token and the max allowance previously given would no
longer be a max allowance, and it would keep reducing for each transferFrom call
breaking the promise previously given.

I have also reworked unit tests a little so that it is clear now when we are
testing max allowance in transferFrom and burnFrom and when we are testing
a never expiring permit.