Compute PERMIT_TYPEHASH constant instead of hardcoding it
Instead of hardcoding the result of keccak256 and explaining in a
comment where does the particular value come from, we compute it
when assigning PERMIT_TYPEHASH value.

I believe the intention behing the original version was to make
deployment a little cheaper, however the compiler should be inteligent
enough to optimize it.