Add a GH Actions workflow publishing npm package with contracts
Contracts stored in the `thesis/solidity-contracts` repository are used
as a dependency in other projects. Currently they are pulled from the
GitHub and referenced to by commit hash.
In this commit we add a GH Actions workflow that will publish npm
packages containing `*.sol` contracts and version them using `-dev.X`
suffix.